Article重构完成

+++ src/main/java/com/kkkoke/networkrepair/dao/ArticleDao.java
@@ -11,13 +11,13 @@ public interface ArticleDao {
     // 添加文章
     int addArticle(Article article);
     // 删除文章
-    int deleteArticle(Long id);
+    int deleteArticle(Integer id);
     // 修改文章
-    Integer updateArticle(Article article);
+    int updateArticle(Article article);
     // 查看所有文章
     List<Article> selectAllArticle();
     // 通过用id查找文章
-    Article selectArticleById(Long id);
+    Article selectArticleById(Integer id);
     // 通过创建时间查找文章
     Article selectArticleByCreateTime(String createTime);
     // 通过修改时间查找文章
